fix all `strict-casts` and `argument_type_not_assignable` (#885)
* password_input_dialog.dart

* mock_ipfs_api.dart line 8

* mock_encointr_api.dart line 108

* search_results_business.dart line 9

* serach_result_offering.dart line 9

* search _result.dart line 46, 47

* business_form_state.dart line 66, 71, 76, 81, 86, 91

* opening_hours.dart line 25, 143

* bazaar_item_data.dart line  24,25, 27, 49, 50, 52

* add_account_page.dart line 113

* index.dart line 551, 446, password_input_dialog.dart line 10

* lib/page/assets/trasfer/payment_confirmation_page/index.dart line 162

* network_select_page.dart line 104

* change_password_page.dart line 84

* voucher.dart line 10

* deep_link.dart line 43, 44, 45, 59

* feed.dart line 16

* http_api.dart line 169, 74

* account_manage_page.dart line 113, 114

* setting_page.dart line 100

* undone

* community_identifier.dart line 50

* account_api.dart line 153, 187

* export_result_page.dart line 16, 76

* reap_voucher_page.dart line 183;

* account_magae.page.dart line 102

* setting_page.dart line 100, encointer.dart line 175, reap_voucher_page.dart line61-66, generated  chain.g.dart

* api.dart line189, 190, 191

* local_storage.dart line 160, 60,194

* encointer_api.dart line 333, 380, 405, 504

* submit_to_js.dart line 65, 91, 109, 110, 11, 113

* line 35, 110

* account.dart line 181, 182, 183, 184, 185, 192, 369

* app.dart line 129

* snack_bar.dart line 17, 30, 43,

* fix all srict-casts

* fix unit test

* fix IOS test

* [MockIpfs] remove non-constant default constructor arg.

* make flutterw executable again

+++ lib/page/assets/transfer/payment_confirmation_page/index.dart
@@ -159,7 +159,7 @@ class _PaymentConfirmationPageState extends State<PaymentConfirmationPage> with 
         _transferState = TransferState.failed;
       } else {
         _transferState = TransferState.finished;
-        _blockTimestamp = DateTime.fromMillisecondsSinceEpoch(res['time']);
+        _blockTimestamp = DateTime.fromMillisecondsSinceEpoch(res['time'] as int);
       }
     };
